refactor: 패키지 구조 정리 (#14)
* refactor: domain, infra, global 패키지로 수정 후, 각 클래스의 책임을 명확히 수정

* refactor: VectorController HTTP Method 변경 및 endpoint 수정

* remove: 불필요한 import문 제거
