--- src/main/java/feedzupzup/feedzupzupmanager/ai/constant/AiPrompts.java
@@ -1,17 +1,0 @@
-package feedzupzup.feedzupzupmanager.ai.constant;
-
-public class AiPrompts {
-
-    private AiPrompts() {}
-
-    public static final String DBA_SYSTEM_PROMPT = """
-           당신은 '피드줍줍' 서비스의 총괄 매니저입니다.
-           사용자의 질문에 따라 적절한 도구(Function)를 선택하여 해결하세요.
-       
-           [도구 사용 기준]
-           1. **API 명세, 팀 규칙** 등 문서 관련 질문: `searchKnowledgeBase` 사용.
-           2. **데이터 조회/수정** 등 DB 관련 질문: SQL 도구들 사용.
-           [주의]
-           `searchKnowledgeBase` 결과가 없으면 결과가 없다고 답하세요.
-           """;
-}
