--- src/main/java/feedzupzup/feedzupzupmanager/ai/dto/QueryLlmRequest.java
@@ -1,5 +1,0 @@
-package feedzupzup.feedzupzupmanager.ai.dto;
-
-public record QueryLlmRequest(String input) {
-
-}
