--- src/main/java/feedzupzup/feedzupzupmanager/ai/dto/QueryLlmResponse.java
@@ -1,5 +1,0 @@
-package feedzupzup.feedzupzupmanager.ai.dto;
-
-public record QueryLlmResponse(String llmResponse) {
-
-}
