+++ src/main/java/feedzupzup/feedzupzupmanager/domain/ai/dto/LlmResponse.java
@@ -1,0 +1,5 @@
+package feedzupzup.feedzupzupmanager.domain.ai.dto;
+
+public record LlmResponse(String llmResponse) {
+
+}
