+++ src/main/java/feedzupzup/feedzupzupmanager/domain/query/exception/NotExecuteQueryException.java
@@ -1,0 +1,13 @@
+package feedzupzup.feedzupzupmanager.domain.query.exception;
+
+import feedzupzup.feedzupzupmanager.global.error.exception.DomainException;
+import feedzupzup.feedzupzupmanager.global.error.ErrorCode;
+
+public class NotExecuteQueryException extends DomainException {
+
+    private static final ErrorCode errorCode = ErrorCode.DANGEROUS_QUERY;
+
+    public NotExecuteQueryException(final String message) {
+        super(errorCode, message);
+    }
+}
