--- src/main/java/feedzupzup/feedzupzupmanager/global/controller/ViewMappingController.java
@@ -1,13 +1,0 @@
-package feedzupzup.feedzupzupmanager.global.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ViewMappingController {
-
-    @GetMapping("/")
-    public String chatPage() {
-        return "chat";
-    }
-}
