--- src/main/java/feedzupzup/feedzupzupmanager/global/exception/DomainException.java
@@ -1,17 +1,0 @@
-package feedzupzup.feedzupzupmanager.global.exception;
-
-import feedzupzup.feedzupzupmanager.global.exception.error.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public abstract class DomainException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-    private final String message;
-
-    protected DomainException(final ErrorCode errorCode, final String message) {
-        super(message);
-        this.message = message;
-        this.errorCode = errorCode;
-    }
-}
