--- src/main/java/feedzupzup/feedzupzupmanager/global/exception/GlobalExceptionHandler.java
@@ -1,62 +1,0 @@
-package feedzupzup.feedzupzupmanager.global.exception;
-
-import feedzupzup.feedzupzupmanager.global.exception.error.ErrorResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.web.HttpRequestMethodNotSupportedException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import static feedzupzup.feedzupzupmanager.global.exception.error.ErrorCode.*;
-
-@RestControllerAdvice
-@Slf4j
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(DomainException.class)
-    public ResponseEntity<ErrorResponse> handleDomainException(final DomainException e) {
-        log.error(e.getMessage(), e);
-        final HttpStatus httpStatus = e.getErrorCode().getHttpStatus();
-        return ResponseEntity.status(httpStatus)
-                .body(ErrorResponse.error(e.getErrorCode()));
-    }
-
-    @ExceptionHandler(DataProcessingException.class)
-    public ResponseEntity<ErrorResponse> handleDataProcessingException(final DataProcessingException e) {
-        log.error(e.getMessage(), e);
-        final HttpStatus httpStatus = e.getErrorCode().getHttpStatus();
-        return ResponseEntity.status(httpStatus)
-                .body(ErrorResponse.error(e.getErrorCode()));
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-    public ResponseEntity<ErrorResponse> handleHttpMessageNotReadableException(final HttpMessageNotReadableException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity.status(INVALID_INPUT_VALUE.getHttpStatus())
-                .body(ErrorResponse.error(INVALID_INPUT_VALUE));
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(final MethodArgumentNotValidException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity.status(INVALID_TYPE_VALUE.getHttpStatus())
-                .body(ErrorResponse.error(INVALID_TYPE_VALUE));
-    }
-
-    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
-    public ResponseEntity<ErrorResponse> handleHttpRequestMethodNotSupportedException(final HttpRequestMethodNotSupportedException e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity.status(METHOD_NOT_ALLOWED.getHttpStatus())
-                .body(ErrorResponse.error(METHOD_NOT_ALLOWED));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleException(final Exception e) {
-        log.error(e.getMessage(), e);
-        return ResponseEntity.status(INTERNAL_SERVER_ERROR.getHttpStatus())
-                .body(ErrorResponse.error(INTERNAL_SERVER_ERROR));
-    }
-}
